feat: Add support for specifying data detectors in FieldContent struct
This commit adds the `data_detector_types` attribute to the `FieldContent` struct in the `ExPass.Structs.FieldContent` module. This attribute allows for specifying a list of data detectors to apply to the field's value, which can automatically convert certain types of data into tappable links. By default, all data detectors are applied. To use no data detectors, an empty list can be specified. This enhancement improves the flexibility and functionality of handling field content in the application.